Refactor sinkmeta structure and device updates
