--- SPONGE/bias/sinkmeta/meta.cpp
@@ -1,189 +1,0 @@
-#include "meta.h"
-
-META::Hill::Hill(const Axis& centers, const Axis& inv_w, const Axis& period,
-                 const float& theight)
-    : height(theight)
-{
-    const int n = static_cast<int>(centers.size());
-    gsf.reserve(n);
-    for (int i = 0; i < n; ++i)
-    {
-        gsf.push_back(GaussianSF(centers[i], inv_w[i], period[i]));
-    }
-    dx_.resize(n);
-    df_.resize(n);
-    tder_.resize(n);
-}
-
-const META::Gdata& META::Hill::Calc_Hill(const Axis& values)
-{
-    const int n = static_cast<int>(values.size());
-    for (int i = 0; i < n; ++i)
-    {
-        dx_[i] = gsf[i].Evaluate(values[i], df_[i]);
-    }
-    potential = 1.0;
-    for (int i = 0; i < n; ++i)
-    {
-        tder_[i] = 1.0;
-        potential *= dx_[i];
-        for (int j = 0; j < n; ++j)
-        {
-            if (j != i)
-            {
-                tder_[i] *= dx_[j];
-            }
-            else
-            {
-                tder_[i] *= df_[j];
-            }
-        }
-    }
-    return tder_;
-}
-
-#ifdef USE_GPU
-static __global__ void Add_Frc(const int atom_numbers, VECTOR* frc,
-                               VECTOR* cv_grad, float dheight_dcv)
-{
-    for (int i = blockIdx.x + blockDim.x * threadIdx.x; i < atom_numbers;
-         i += gridDim.x * blockDim.x)
-    {
-        frc[i] = frc[i] - dheight_dcv * cv_grad[i];
-    }
-}
-
-static __global__ void Add_Potential_Kernel(float* d_potential, const float to_add)
-{
-    d_potential[0] += to_add;
-}
-
-static __global__ void Add_Virial(LTMatrix3* d_virial, const float dU_dCV,
-                                  const LTMatrix3* cv_virial)
-{
-    d_virial[0] = d_virial[0] - dU_dCV * cv_virial[0];
-}
-#else
-static void Add_Frc(const int atom_numbers, VECTOR* frc, VECTOR* cv_grad,
-                    float dheight_dcv)
-{
-#pragma omp parallel for
-    for (int i = 0; i < atom_numbers; i++)
-    {
-        frc[i] = frc[i] - dheight_dcv * cv_grad[i];
-    }
-}
-
-static void Add_Potential_Kernel(float* d_potential, const float to_add)
-{
-    d_potential[0] += to_add;
-}
-
-static void Add_Virial(LTMatrix3* d_virial, const float dU_dCV,
-                       const LTMatrix3* cv_virial)
-{
-    d_virial[0] = d_virial[0] - dU_dCV * cv_virial[0];
-}
-#endif
-
-void META::Meta_Force_With_Energy_And_Virial(int atom_numbers, VECTOR* frc,
-                                             int need_potential,
-                                             int need_pressure,
-                                             float* d_potential,
-                                             LTMatrix3* d_virial)
-{
-    if (!is_initialized)
-    {
-        return;
-    }
-    Potential_And_Derivative(need_potential);
-    if (do_borderwall)
-    {
-        Border_Derivative(border_upper.data(), border_lower.data(), cutoff,
-                          Dpotential_local);
-    }
-
-    for (int i = 0; i < cvs.size(); ++i)
-    {
-        Launch_Device_Kernel(Add_Frc, (atom_numbers + 31 / 32), 32, 0, NULL,
-                             atom_numbers, frc, cvs[i]->crd_grads,
-                             Dpotential_local[i]);
-        if (need_pressure)
-        {
-            Launch_Device_Kernel(Add_Virial, 1, 1, 0, NULL, d_virial,
-                                 Dpotential_local[i], cvs[i]->virial);
-        }
-    }
-    if (need_potential)
-    {
-        Launch_Device_Kernel(Add_Potential_Kernel, 1, 1, 0, NULL, d_potential,
-                             potential_local);
-    }
-}
-
-void META::Potential_And_Derivative(const int need_potential)
-{
-    if (!is_initialized)
-    {
-        return;
-    }
-    for (int i = 0; i < cvs.size(); ++i)
-    {
-        est_values_[i] = cvs[i]->value;
-        Dpotential_local[i] = 0.f;
-    }
-    Estimate(est_values_, need_potential, true);
-}
-
-void META::Border_Derivative(float* border_upper, float* border_lower,
-                             float* cutoff, float* Dpotential_local)
-{
-    for (int i = 0; i < cvs.size(); ++i)
-    {
-        float h_cv = cvs[i]->value;
-        if (h_cv - border_lower[i] < cutoff[i])
-        {
-            float distance = border_lower[i] - h_cv;
-            if (periods[i] > 0)
-            {
-                distance -= roundf(distance / cv_periods[i]) * cv_periods[i];
-            }
-            Dpotential_local[i] =
-                Dpotential_local[i] - border_potential_height * expf(distance);
-        }
-        else if (border_upper[i] - h_cv < cutoff[i])
-        {
-            float distance = h_cv - border_upper[i];
-            if (periods[i] > 0)
-            {
-                distance -= roundf(distance / cv_periods[i]) * cv_periods[i];
-            }
-            Dpotential_local[i] =
-                Dpotential_local[i] + border_potential_height * expf(distance);
-        }
-    }
-}
-
-void META::Do_Metadynamics(int atom_numbers, VECTOR* crd, LTMatrix3 cell,
-                           LTMatrix3 rcell, int step, int need_potential,
-                           int need_pressure, VECTOR* frc, float* d_potential,
-                           LTMatrix3* d_virial, float sys_temp)
-{
-    if (this->is_initialized)
-    {
-        int need = CV_NEED_GPU_VALUE | CV_NEED_CRD_GRADS;
-        if (need_pressure)
-        {
-            need |= CV_NEED_VIRIAL;
-        }
-
-        for (int i = 0; i < cvs.size(); i = i + 1)
-        {
-            this->cvs[i]->Compute(atom_numbers, crd, cell, rcell, need, step);
-        }
-        temperature = sys_temp;
-        Meta_Force_With_Energy_And_Virial(atom_numbers, frc, need_potential,
-                                          need_pressure, d_potential, d_virial);
-        Add_Potential(sys_temp, step);
-    }
-}
